Delete tarantula not working if imgName==null (if statement not helping?) delete account from menu crashing app(AlertDialog?)

+++ app/src/main/java/vloboda/tarantulaProject/myapplication/Tarantulas.java
@@ -118,10 +118,11 @@ public class Tarantulas extends AppCompatActivity {
                     public void onClick(DialogInterface dialog, int which) {
 
                        tarantulaArrayList.forEach((tarantula) ->{
-                           FirebaseStorage.getInstance().getReference()
-                                   .child("tarantulas")
-                                   .child(tarantula.imgName).delete();
-
+                           if(!tarantula.imgName.isEmpty()) {
+                               FirebaseStorage.getInstance().getReference()
+                                       .child("tarantulas")
+                                       .child(tarantula.imgName).delete();
+                           }
                         });
                        FirebaseFirestore.getInstance().collection("users")
                                .document(FirebaseAuth.getInstance().getUid().toString());
